backend/djangoAirbnb_backend/requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
dj-rest-auth==4.0.0
django-allauth==0.53.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-filter==23.5
pillow==10.2.0
channels==4.0.0
daphne==4.0.0
python-dotenv==1.0.0

